- Diverging from our initial idea of fetching videos from youtube with callbacks, we decided to imple ment the connections between different view renderables and model renderables first. This is the init ial version for the mediaplayer. - Leaving Main Activity blank and launching PhotosActivity when the program is first run. - Using an arfragment instead of an ArSceneView. - Starting the mediaplayer with touch events via a tap event listener on the fragment and launching t he mediaplayer. Setting it as the texture to the chroma_key video player obj file.
