--- app/src/main/res/layout/test_view.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<fragment android:id="@+id/youtubeFragment"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:name="com.google.android.youtube.player.YouTubePlayerFragment"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-</fragment>
-
